break up cli script
